requirements.txt
python-telegram-bot==21.8
aiohttp==3.10.11
python-dotenv==1.0.1
nest-asyncio==1.6.0